+++ sarif/operations/blame_op.py
@@ -6,6 +6,8 @@ import json
 import os
 import subprocess
 import sys
+import urllib.parse
+import urllib.request
 
 from sarif.sarif_file import SarifFileSet
 
@@ -94,10 +96,20 @@ def _enhance_with_blame(input_files, repo_path):
     print(f"Found blame information for {blame_info_count} of {item_count} results")
 
 
+def _make_path_git_compatible(file_path):
+    try:
+        path_as_url = urllib.parse.urlparse(file_path)
+        if path_as_url.scheme == "file":
+            return urllib.request.url2pathname(path_as_url.path)
+        return file_path
+    except ValueError:
+        return file_path
+
+
 def _run_git_blame_on_files(files_to_blame, repo_path):
     file_blame_info = {}
     for file_path in files_to_blame:
-        cmd = ["git", "blame", "--porcelain", file_path]
+        cmd = ["git", "blame", "--porcelain", _make_path_git_compatible(file_path)]
         with subprocess.Popen(cmd, stdout=subprocess.PIPE, cwd=repo_path) as proc:
             blame_info = {"commits": {}, "line_to_commit": {}}
             file_blame_info[file_path] = blame_info
@@ -131,6 +143,7 @@ def _run_git_blame_on_files(files_to_blame, repo_path):
             if proc.returncode:
                 cmd_str = " ".join(cmd)
                 sys.stderr.write(
-                    f"WARNING: Command `{cmd_str}` failed with exit code {proc.returncode} in {repo_path}\n"
+                    f"WARNING: Command `{cmd_str} "
+                    f"failed with exit code {proc.returncode} in {repo_path}\n"
                 )
     return file_blame_info